Move exportLogs to LogUtils
Also adds toast messages to it
